# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kadane Backend API
  description: Kadane Backend API 
  version: 1.0.0
servers:
  - url: https://api.kadane.xyz/v1
    description: Main production server
  - url: http://localhost:3000/v1
    description: Local development server
security: 
  -  Authorization: []
paths:
...
  /accounts/username/{username}:
    get:
      tags:
        - Accounts
      summary: Get an account by username
      description: Get an account by username.
      operationId: getAccountByUsername
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: The username of the account to get.
      responses:
        '200':
          description: Account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /friends:
    get:
      tags:
        - Friends
      summary: Get all friends
      description: Get all friends for a user.
      operationId: getFriends
      responses:
        '200':
          description: A list of friends
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FriendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Friends
      summary: Create a friend request
      description: Create a friend request for a user.
      operationId: createFriendRequest
      requestBody:
        description: Friend request object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
      responses:
        '201':
          description: Friend request created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /friends/{username}:
    delete:
      tags:
        - Friends
      summary: Delete a friend
      description: Delete a friend by their username.
      operationId: deleteFriend
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: The username of the friend to delete.
      responses:
        '204':
          description: Friend deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /friends/username/{username}:
    get:
      tags:
        - Friends
      summary: Get friends by username
      description: Get friends by username.
      operationId: getFriendsByUsername
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: The username of the friend to get.
      responses:
        '200':
          description: A list of friends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /friends/requests/accept:
    post:
      tags:
        - Friends
      summary: Accept a friend request
      description: Accept a friend request for a user.
      operationId: acceptFriendRequest
      requestBody:
        description: Friend request object to be accepted
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
      responses:
        '201':
          description: Friend request accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /friends/requests/block:
    post:
      tags:
        - Friends
      summary: Block a friend request
      description: Block a friend request for a user.
      operationId: blockFriendRequest
      requestBody:
        description: Friend request object to be blocked
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
      responses:
        '201':
          description: Friend request blocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendResponse'
  /friends/requests/unblock:
    post:
      tags:
        - Friends
      summary: Unblock a friend request
      description: Unblock a friend request for a user.
      operationId: unblockFriendRequest
      requestBody:
        description: Friend request object to be unblocked
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
      responses:
        '201':
          description: Friend request unblocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /friends/requests/{username}:
    delete:
      tags:
        - Friends
      summary: Delete a friend request
      description: Delete a friend request by username.
      operationId: deleteFriendRequest
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: The username of the friend request to delete.
      responses:
        '204':
          description: Friend request deleted successfully
...
components:
  securitySchemes: 
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: jwt 
  schemas:
...
    StatusCode:
      type: string
      enum: [200, 201, 202, 204, 400, 401, 403, 404, 422, 500]
    
    # Responses
    # APIResponse is a generic response that can be used for any response
    # { "data": { any } }
...
    APIErrorData:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
          anyOf:
            - $ref: '#/components/schemas/StatusCode'
    APIError:
      allOf:
        - $ref: '#/components/schemas/APIErrorData'
        - type: object
          properties:
            error:
              type: object
              $ref: '#/components/schemas/APIErrorData'
...
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/APIError'
        - type: object
          properties:
            code:
              type: string
              enum: [500]

    # Accounts
    Account:
      allOf:
        - type: object
          properties:
            id:
              type: string
            username:
              type: string
            email:
              type: string
              format: email
            avatarUrl:
              type: string
            level:
              type: integer
            createdAt:
              type: string
              format: date-time
            friendStatus:
              type: string
              enum: [none, friend, blocked, request_sent, request_received]
            plan:
              type: string
              enum: [free, plus, pro]
            isAdmin:
              type: boolean
            attributes:
              $ref: '#/components/schemas/AccountAttributes'
          required:
            - id
            - username
            - email
    AccountAttributes:
      description: Account attributes
      type: object
      properties:
        bio:
          type: string
          nullable: true
        contactEmail:
          type: string
          format: email
          nullable: true
        location:
          type: string
          nullable: true
        realName:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        facebookUrl:
          type: string
          nullable: true
        instagramUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        school:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        friendCount:
          type: integer
        blockedCount:
          type: integer
        friendRequestCount:
          type: integer
...
    AccountResponse:
      description: Account response
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Account'
...
    Friend:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
        location:
          type: string
        level:
          type: integer
        acceptedAt:
          type: string
          format: date-time
    FriendRequest:
      type: object
      properties:
        friendName:
          type: string
    FriendResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Friend'
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'